Adjust test expectation to address semver hazards failure
